Fix patch on _run_*_callbacks
Rails 4.2.3+ had reverted the commit that changed the private callbacks
API, see
https://github.com/rails/rails/commit/fe664e8ce97b5f7a1e1d95dd4df0fbdba42c7f2b.
Thus, we cannot rely on the Rails version, as 4.2.4 was not released
yet. I changed the check to check against a private method that
existed/dont before and after the patch.
Such a hack =/ , but unfortunately we are dealing with internal apis
that can change at any time.