Remove deprecated methods
Closes #13

The primary use for this gem is from the identity cache gem and it
doesn't need the deprecated behaviour.  With rails 5.1 identity
cache will now get a lot of warnings when this deprecated behaviour
is included in a model, even when it isn't used.

Rather than supporting these deprecated methods in this gem, they can
always be extracted by any app that actually needs them into a separate
module to be included where needed (e.g. DeprecatedArTransactionChanges).
So I am going to remove them before making another release.